requirements.txt
psycopg[binary]>=3.1.18
pytest>=8.0.0
python-dotenv>=1.0.0
